requirements.txt
streamlit>=1.35
pydantic>=2.6
pyyaml>=6.0
openai>=1.40.0
